requirements.txt
python-whois==0.8.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
dnspython==2.4.2



pytest>=8.0.0
